+++ CHANGELOG.md
@@ -15,6 +15,10 @@ You can also check the
 
   - Added Overview Details Page
 
+Enhancements
+
+- Add Query state to details pages
+
 # 2.3.1 - 2025-06-26
 
 - Fixes